pyproject.toml
[project]
name = "il-news-bot"
version = "0.1.0"
description = "An agentic bot to get non-biased news"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "asyncio>=4.0.0",
  "feedparser>=6.0.12",
  "langchain>=1.2.15",
  "langchain-core>=1.3.0",
  "langchain-openrouter>=0.2.1",
  "langgraph>=1.1.8",
  "llm-guard>=0.3.16",
  "pydantic>=2.13.2",
  "python-dotenv>=1.2.2",
  "python-telegram-bot>=22.7",
  "telethon>=1.43.2",
]

[dependency-groups]
dev = ["ruff>=0.15.11", "ty>=0.0.31"]

[tool.ruff]
line-length = 88

[tool.ruff.format]
indent-style = "tab"

[tool.ruff.lint]
extend-select = ["E501"]
